refactor: inicia a reestruturação do projeto para um editor de código na web simplificado
* Substitui o estado de domínio (quests, skills, avatars) por estruturas de programação (variables, functions, nodes).
* Introduz a interface SDKNode e o sistema de conexões por execução/dados.
* Remove modelos obsoletos de RPG.
