--- apps/web/src/lib/editor/actions.ts
@@ -1,148 +1,0 @@
-import { computed } from "vue";
-import { createDefinitions } from "@projeto/types";
-import { generateId, useEditorStore } from "@/stores/editor.ts";
-
-export const actions = computed(() => {
-	const editorStore = useEditorStore();
-
-	function pushAction(type: string, params: any) {
-		const id = generateId("action");
-		const action = { id, type, ...params };
-		editorStore.actions.push(action);
-		return id;
-	}
-
-	const definitions = createDefinitions({
-		showDialogue: {
-			text: "Mostrar Diálogo",
-			icon: "chat",
-			params: [
-				{
-					key: "speaker",
-					label: "Nome do Orador",
-					type: "text",
-					required: true,
-				},
-				{
-					key: "text",
-					label: "Texto do Diálogo",
-					type: "textarea",
-					required: true,
-				},
-				{
-					key: "avatarId",
-					label: "Avatar (Opcional)",
-					type: "select",
-					items: editorStore.avatars,
-				},
-			],
-			execute: (params) => pushAction("showDialogue", params),
-		},
-		hideUI: {
-			text: "Esconder UI",
-			icon: "visibility_off",
-			params: [
-				{
-					key: "element",
-					label: "Elemento da UI",
-					type: "select",
-					items: ["all", "dialogue_box", "hotbar", "hud"],
-					default: "all",
-				},
-			],
-			execute: (params) => pushAction("hideUI", params),
-		},
-		showUI: {
-			text: "Mostrar UI",
-			icon: "visibility",
-			params: [
-				{
-					key: "element",
-					label: "Elemento da UI",
-					type: "select",
-					items: ["all", "dialogue_box", "hotbar", "hud"],
-					default: "all",
-				},
-			],
-			execute: (params) => pushAction("showUI", params),
-		},
-		moveCamera: {
-			text: "Mover Câmera",
-			icon: "videocam",
-			params: [
-				{ key: "x", label: "Posição X", type: "number", required: true },
-				{ key: "y", label: "Posição Y", type: "number", required: true },
-				{
-					key: "duration",
-					label: "Duração (s)",
-					type: "number",
-					required: true,
-				},
-			],
-			execute: (params) => pushAction("moveCamera", params),
-		},
-		animateObject: {
-			text: "Animar Objeto",
-			icon: "animation",
-			params: [
-				{
-					key: "targetId",
-					label: "Alvo",
-					type: "select",
-					items: editorStore.objects,
-					required: true,
-				},
-				{
-					key: "animationName",
-					label: "Nome da Animação",
-					type: "text",
-					required: true,
-				},
-			],
-			execute: (params) => pushAction("animateObject", params),
-		},
-		wait: {
-			text: "Aguardar",
-			icon: "pause",
-			params: [
-				{
-					key: "duration",
-					label: "Duração (s)",
-					type: "number",
-					required: true,
-				},
-			],
-			execute: (params) => pushAction("wait", params),
-		},
-		setBackgroundImage: {
-			text: "Mudar Fundo",
-			icon: "image",
-			params: [
-				{ key: "url", label: "URL da Imagem", type: "text", required: true },
-			],
-			execute: (params) => pushAction("setBackgroundImage", params),
-		},
-		setMusic: {
-			text: "Tocar Música",
-			icon: "music_note",
-			params: [
-				{ key: "url", label: "URL da Música", type: "text", required: true },
-			],
-			execute: (params) => pushAction("setMusic", params),
-		},
-		setSoundEffect: {
-			text: "Efeito Sonoro",
-			icon: "volume_up",
-			params: [
-				{ key: "url", label: "URL do Efeito", type: "text", required: true },
-			],
-			execute: (params) => pushAction("setSoundEffect", params),
-		},
-	});
-
-	return {
-		text: "Ação",
-		icon: "https://fonts.gstatic.com/s/e/notoemoji/latest/1f3ac/lottie.json",
-		definitions,
-	};
-});
