--- packages/types/domain/editor/index.ts
@@ -1,2 +1,0 @@
-export * from "./config";
-export * from "./models";
