--- packages/types/domain/index.ts
@@ -1,3 +1,0 @@
-export * from "./session";
-export * from "./user";
-export * from "./editor"
