--- packages/types/domain/user.ts
@@ -1,10 +1,0 @@
-export interface BackendUserAuth {
-	accessToken: string;
-	refreshToken: string;
-}
-
-export interface BackendUser extends BackendUserAuth {
-	id: string;
-	email: string;
-	name: string;
-}
