+++ packages/types/editor/tokens.ts
@@ -1,0 +1,30 @@
+/**
+ * Palavras reservadas
+ * **/
+export const Keywords = {
+	LET: "let",
+	CONST: "const",
+	IF: "if",
+	ELSE: "else",
+	FUNCTION: "function",
+	RETURN: "return",
+} as const;
+
+/**
+ * Operadores aritméticos, lógicos e de comparação
+ * **/
+export const Operators = {
+	ADD: "+",
+	SUB: "-",
+	MUL: "*",
+	DIV: "/",
+	EQ: "===",
+	NEQ: "!==",
+	GT: ">",
+	LT: "<",
+	GTE: ">=",
+	LTE: "<=",
+	AND: "&&",
+	OR: "||",
+	NOT: "!",
+} as const;
